Chore: Notification Name 위치 수정 - post 선언된 위치에 해당 Notification Name 선언
