--- DrawingApp/View/Notification/NotificationName.swift
@@ -1,18 +1,0 @@
-//
-//  RightAttributerViewDelegate.swift
-//  DrawingApp
-//
-//  Created by juntaek.oh on 2022/03/03.
-//
-
-import Foundation
-
-extension Notification.Name{
-    static let makeRectangle = Notification.Name("makeRectangle")
-    static let tapGesture = Notification.Name("tapGesture")
-    
-    static let changeAlpha = Notification.Name("changeAlpha")
-    static let changeRedColor = Notification.Name("changeRedColor")
-    static let changeGreenColor = Notification.Name("changeGreenColor")
-    static let changeBlueColor = Notification.Name("changeBlueColor")
-}
