+++ DrawingApp/View/RightAttributerView.swift
@@ -169,3 +169,10 @@ extension RightAttributerView{
         NotificationCenter.default.post(name: .changeBlueColor, object: nil)
     }
 }
+
+extension Notification.Name{
+    static let changeAlpha = Notification.Name("changeAlpha")
+    static let changeRedColor = Notification.Name("changeRedColor")
+    static let changeGreenColor = Notification.Name("changeGreenColor")
+    static let changeBlueColor = Notification.Name("changeBlueColor")
+}
